domain/video.go
package domain

import (
	"time"

	"github.com/asaskevich/govalidator"
)

type Video struct {
	ID         string    `json:"encoded_video_folder" valid:"uuid" gorm:"type:uuid;primary_key"`
	ResourceId string    `json:"resource_id" valid:"notnull" gorm:"type:varchar"`
	FilePath   string    `json:"file_path" valid:"notnull" gorm:"type:varchar"`
	CreatedAt  time.Time `json:"-" valid:"-"`
	UpdatedAt  time.Time `json:"-" valid:"-"`
	Jobs       []*Job    `json:"-" valid:"-" gorm:"ForeingKey:VideoId"`
}

func init() {
	govalidator.SetFieldsRequiredByDefault(true)
}

func NewVideo() *Video {
	return &Video{}
}

func (video *Video) Validate() error {
	_, err := govalidator.ValidateStruct(video)

	if err != nil {
		return err
	}

	return nil
}
